Moved to SQLite

+++ stroom-proxy/stroom-proxy-repo/src/main/java/stroom/proxy/repo/ZipInfoStore.java
@@ -20,5 +20,5 @@ import java.nio.file.Path;
 
 public interface ZipInfoStore {
 
-    Long store(Path path, Path relativePath, ErrorReceiver errorReceiver);
+    int store(Path path, Path relativePath, ErrorReceiver errorReceiver);
 }
